Using go:embed to embed console into the go binary
